Implement registration valid form
